+++ clients/admin/src/lib/salesInvoices.ts
@@ -1,6 +1,6 @@
-import gql from 'graphql-tag';
 import { query } from 'svelte-apollo';
 import type { SalesInvoiceByIdQuery } from '../generated/graphql';
+import { GET_SALES_INVOICE_BY_ID } from './queries/salesInvoice';
 
 export const downloadInvoice = (
     baseUrl: string | undefined,
@@ -21,97 +21,6 @@ export const downloadInvoice = (
             a.click();
         });
 };
-
-const GET_SALES_INVOICE_BY_ID = gql`
-    query SalesInvoiceById($id: Int!) {
-        salesInvoice(id: $id) {
-            currency {
-                id
-                isoCode
-                displayName
-            }
-            customer {
-                id
-                legalName
-                displayName
-                vatNumber
-                invoicingEmail
-                legalAddress {
-                    id
-                    city
-                    line1
-                    zipCode
-                    country {
-                        id
-                        isoCode
-                    }
-                }
-                address {
-                    id
-                    city
-                    line1
-                    zipCode
-                    country {
-                        id
-                        isoCode
-                    }
-                }
-                note
-            }
-            documentNo
-            dueDate
-            grandTotal
-            grandTotalAccountingSchemeCurrency
-            id
-            isActive
-            isCalculated
-            isCurrent
-            isDraft
-            issuedOn
-            lines {
-                id
-                lineOrder
-                linePrice
-                narration
-                quantity
-                product {
-                    id
-                }
-            }
-            organization {
-                contact
-                displayName
-                id
-                idNumber
-                legalName
-                registration
-                vatNumber
-                accountingScheme {
-                    currency {
-                        displayName
-                    }
-                }
-            }
-            paymentTermInDays
-            printDate
-            printed
-            printError
-            printLanguageIsoCode
-            reverseCharge
-            totalLines
-            totalLinesAccountingSchemeCurrency
-            transactionDate
-            vatReport {
-                id
-                vatRatePercent
-                vatTotal
-                vatTotalAccountingSchemeCurrency
-                vatTotalAccountingSchemeCurrencyRaw
-                vatTotalRaw
-            }
-        }
-    }
-`;
 
 export const getSalesInvoiceBy = (id: number) =>
     query<SalesInvoiceByIdQuery>(GET_SALES_INVOICE_BY_ID, { variables: { id } });
